Pvp Combat feature (#678)

+++ src/GameWorldSimulator/NeoServer.Game.Creatures/Models/Bases/WalkableCreature.cs
@@ -286,7 +286,7 @@ public abstract class WalkableCreature : Creature, IWalkableCreature
         if (_walkingQueue.TryDequeue(out direction))
         {
             FirstStep = false;
-            Cooldowns.Start(CooldownType.Move, StepDelay);
+            Cooldowns.Start(CooldownType.Move, (uint)StepDelay);
 
             return true;
         }
